pyproject.toml
[project]
name = "linguado"
version = "0.2.0"
description = "Linguado is a tool which compares the AST of two or more files"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
keywords = ["AST", "compare"]
authors = [
  {name = "Guzmán Cernadas Pérez", email = "guzman.cernadas@protonmail.com"},
]
maintainers = [
  {name = "Guzmán Cernadas Pérez", email = "guzman.cernadas@protonmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]

dependencies = [
    "numpy",
    "tqdm",
    "igraph",
    "networkx",
    "antlr4-python3-runtime",
    "scipy",
    "setuptools"
]

[project.urls]
repository = "https://github.com/Caralludo/linguado.git"

[project.scripts]
linguado = "linguado:main"
